+++ mapsforge-poi-writer/src/main/java/org/mapsforge/poi/writer/PoiWriter.java
@@ -162,7 +162,9 @@ public final class PoiWriter {
 
         try {
             commit();
-            filterCategories();
+            if (configuration.isFilterCategories()) {
+                filterCategories();
+            }
             writeMetadata();
             this.conn.close();